ci: split ubuntu and Macos builds
As a preparation to use a custom container image for Ubuntu.

Currently, each Ubuntu runner takes more than 8 minutes to complete.
That's because they spent a lot of time installing dependencies,
in particular all lcov's Perl dependencies.

It's better to use a container runner for Ubuntu and prepare
it separately before running the actual build of the C++ project.